feat: Add settings page and integrate with authentication for user profile management
